Add settings for permissions, enhance audio upload management, and improve search functionality
- Introduced settings.json for permission management.
- Updated LibraryUploadsApi to handle upload limits correctly, ensuring no invalid arguments are thrown.
- Enhanced LibraryUploadsNotifier to remove deleted tracks from the active queue.
- Improved public user uploads provider to handle backend changes in track data structure.
- Updated SearchNotifier to inject private tracks into search results for better user experience.
- Added functionality to PlayerNotifierQueue to remove tracks by ID from the queue.
- Modified PlayerWaveformBar to support fractional seeking.
- Updated TrackOptionsSheet to manage sharing permissions based on track ownership and privacy.
- Adjusted tests to accommodate changes in data types for seeking functionality.

+++ lib/features/playback_streaming_engine/presentation/screens/player_screen_body.dart
@@ -238,7 +238,7 @@ class _TrackContent extends ConsumerWidget {
 
   final PlayerState playerState;
   final Animation<double> artworkScale;
-  final void Function(int positionSeconds) onSeek;
+  final void Function(double positionSeconds) onSeek;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
